+++ apps/e2e/tests/backend/endpoints/api/v1/auth/otp/send-sign-in-code.test.ts
@@ -8,7 +8,7 @@ it("should send a sign-in code per e-mail", async ({ expect }) => {
       MailboxMessage {
         "from": "Stack Dashboard <noreply@example.com>",
         "subject": "Sign in to Stack Dashboard",
-        "to": ["<<stripped UUID>@stack-generated.example.com>"],
+        "to": ["<default-mailbox--<stripped UUID>@stack-generated.example.com>"],
         <some fields may have been hidden>,
       },
     ]