+++ apps/e2e/tests/backend/endpoints/api/v1/projects.test.ts
@@ -1,5 +1,5 @@
 import { isBase64Url } from "@stackframe/stack-shared/dist/utils/bytes";
-import { createMailbox, it } from "../../../../helpers";
+import { it } from "../../../../helpers";
 import { Auth, InternalProjectKeys, Project, backendContext, niceBackendFetch } from "../../../backend-helpers";
 
 
@@ -1005,14 +1005,11 @@ it("deletes a project with users, teams, and permissions", async ({ expect }) =>
 });
 
 it("makes sure user have the correct managed project ID after project creation", async ({ expect }) => {
-  const { userId } =  await Auth.Otp.signIn();
+  backendContext.set({ projectKeys: InternalProjectKeys });
+  const { creatorUserId, projectId } = await Project.createAndGetAdminToken();
 
   backendContext.set({ projectKeys: InternalProjectKeys });
-  const { projectId } = await Project.createAndGetAdminToken();
-
-  backendContext.set({ projectKeys: InternalProjectKeys });
-
-  const userResponse = await niceBackendFetch(`/api/v1/users/${userId}`, {
+  const userResponse = await niceBackendFetch(`/api/v1/users/${creatorUserId}`, {
     accessType: "server",
     method: "GET",
   });
@@ -1022,10 +1019,8 @@ it("makes sure user have the correct managed project ID after project creation",
 });
 
 it("makes sure user don't have managed project ID after project deletion", async ({ expect }) => {
-  const { userId } =  await Auth.Otp.signIn();
-
   backendContext.set({ projectKeys: InternalProjectKeys });
-  const { adminAccessToken } = await Project.createAndGetAdminToken();
+  const { creatorUserId, adminAccessToken } = await Project.createAndGetAdminToken();
 
   // Delete the project
   const deleteResponse = await niceBackendFetch(`/api/v1/projects/current`, {
@@ -1046,7 +1041,7 @@ it("makes sure user don't have managed project ID after project deletion", async
 
   backendContext.set({ projectKeys: InternalProjectKeys });
 
-  const userResponse = await niceBackendFetch(`/api/v1/users/${userId}`, {
+  const userResponse = await niceBackendFetch(`/api/v1/users/${creatorUserId}`, {
     accessType: "server",
     method: "GET",
   });
@@ -1055,12 +1050,9 @@ it("makes sure user don't have managed project ID after project deletion", async
 });
 
 it("makes sure other users are not affected by project deletion", async ({ expect }) => {
-  const { userId: userId1 } =  await Auth.Otp.signIn();
   backendContext.set({ projectKeys: InternalProjectKeys });
-  const { projectId } = await Project.createAndGetAdminToken();
+  const { creatorUserId, projectId } = await Project.createAndGetAdminToken();
 
-  backendContext.set({ mailbox: createMailbox(), projectKeys: InternalProjectKeys });
-  await Auth.Otp.signIn();
   backendContext.set({ projectKeys: InternalProjectKeys });
   const { adminAccessToken } = await Project.createAndGetAdminToken();
 
@@ -1074,7 +1066,7 @@ it("makes sure other users are not affected by project deletion", async ({ expec
   });
 
   backendContext.set({ projectKeys: InternalProjectKeys });
-  const userResponse1 = await niceBackendFetch(`/api/v1/users/${userId1}`, {
+  const userResponse1 = await niceBackendFetch(`/api/v1/users/${creatorUserId}`, {
     accessType: "server",
     method: "GET",
   });