+++ apps/dev-launchpad/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @stackframe/dev-launchpad
 
+## 2.7.0
+
 ## 2.6.39
 
 ## 2.6.38