+++ examples/demo/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @stackframe/example-demo-app
 
+## 2.7.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.7.0
+  - @stackframe/stack-ui@2.7.0
+  - @stackframe/stack@2.7.0
+
 ## 2.6.39
 
 ### Patch Changes