+++ examples/supabase/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @stackframe/example-supabase
 
+## 2.7.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack@2.7.0
+
 ## 2.6.39
 
 ### Patch Changes